requirements.txt
fastapi
python-multipart
uvicorn
python-dotenv
pillow
fal-client
requests
